move code-mirror selection functions into selections.ts
